feat(*): upgrade ahlib-more, adjust some codes of xgin
