+++ README.md
@@ -1,6 +1,12 @@
 # ahlib-web
 
-+ A personal golang library for personal web development, including gin + bot, split from [Aoi-hosizora/ahlib](https://github.com/Aoi-hosizora/ahlib).
+[![Build Status](https://travis-ci.com/Aoi-hosizora/ahlib-web.svg?branch=master)](https://travis-ci.com/Aoi-hosizora/ahlib-web)
+[![codecov](https://codecov.io/gh/Aoi-hosizora/ahlib-web/branch/master/graph/badge.svg)](https://codecov.io/gh/Aoi-hosizora/ahlib-web)
+[![Go Report Card](https://goreportcard.com/badge/github.com/Aoi-hosizora/ahlib-web)](https://goreportcard.com/report/github.com/Aoi-hosizora/ahlib-web)
+[![License](http://img.shields.io/badge/license-mit-blue.svg)](./LICENSE)
+[![Release](https://img.shields.io/github/v/release/Aoi-hosizora/ahlib-web)](https://github.com/Aoi-hosizora/ahlib-web/releases)
+
++ A personal golang library for web development, including gin, telebot, serverchan.
 
 ### Related libraries
 
@@ -20,8 +26,11 @@
 
 + See [go.mod](./go.mod) and [go.sum](./go.sum)
 + `github.com/Aoi-hosizora/ahlib v1.5.0`
++ `github.com/Aoi-hosizora/ahlib-more v1.5.0`
 + `github.com/gin-gonic/gin v1.6.3`
 + `github.com/go-playground/validator/v10 v10.2.0`
++ `github.com/go-playground/universal-translator v0.17.0`
++ `github.com/go-playground/locales v0.13.0`
 + `gopkg.in/tucnak/telebot.v2 v2.3.3`
 + `github.com/Aoi-hosizora/go-serverchan v1.0.1`
 + `github.com/sirupsen/logrus v1.7.0`
