+++ go.mod
@@ -4,6 +4,7 @@ go 1.13
 
 require (
 	github.com/Aoi-hosizora/ahlib v1.5.0
+	github.com/Aoi-hosizora/ahlib-more v1.5.0
 	github.com/Aoi-hosizora/go-serverchan v1.0.1
 	github.com/gin-gonic/gin v1.6.3
 	github.com/go-playground/locales v0.13.0
